+++ src/Symfony/Bundle/TwigBundle/Tests/DependencyInjection/TwigExtensionTest.php
@@ -32,7 +32,7 @@ class TwigExtensionTest extends TestCase
         $container->registerExtension(new TwigExtension());
         $container->loadFromExtension('twig', [
             'strict_variables' => false, // to be removed in 5.0 relying on default
-            'exception_controller' => null, // to be removed in 5.0 relying on default
+            'exception_controller' => null, // to be removed in 5.0
         ]);
         $this->compileContainer($container);