+++ src/Symfony/Bundle/FrameworkBundle/CHANGELOG.md
@@ -14,7 +14,8 @@ CHANGELOG
  * Deprecated `routing.loader.service`, use `routing.loader.container` instead.
  * Not tagging service route loaders with `routing.route_loader` has been deprecated.
  * Overriding the methods `KernelTestCase::tearDown()` and `WebTestCase::tearDown()` without the `void` return-type is deprecated.
- 
+ * Added new `error_controller` configuration to handle system exceptions
+
 4.3.0
 -----